requirements.txt
Flask==3.0.0
loguru==0.7.2
Werkzeug==3.0.1
gunicorn==21.2.0
PyJWT==2.8.0

